requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.0
distro==1.9.0
executing==2.2.0
fastapi==0.115.8
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipython==8.32.0
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.37
langchain-google-genai==2.0.10
langchain-openai==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.10
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
openai==1.63.2
orjson==3.10.15
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyparsing==3.2.1
python-dotenv==1.0.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
zstandard==0.23.0
